pyproject.toml
[project]
name = "air-marshall"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = []

[project.scripts]
air-marshall-db = "air_marshall.database.__main__:main"

[project.optional-dependencies]
hvac = []      # RPi deps for HVAC controller (e.g. RPi.GPIO, pimoroni HAT libs)
monitor = []   # RPi deps for monitor (e.g. sensor libraries)
database = [
    "fastapi>=0.111",
    "uvicorn[standard]>=0.29",
    "aiosqlite>=0.20",
    "pydantic-settings>=2.2",
    "httpx>=0.27",
]
test = [
    "pytest>=8.0",
    "pytest-cov>=4.1",
    "pytest-asyncio>=0.23",
    "pytest-timeout>=2.1",
]
lint = [
    "ruff>=0.1",
    "mypy>=1.6",
]
dev = ["air-marshall[hvac,monitor,database,test,lint]"]

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.uv]
package = true

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "D", "S", "C4"]
ignore = ["E501", "D100", "D104", "D107", "B008"]  # B008: FastAPI Depends/Query as default args is intentional

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S101", "D103"]  # assert ok in tests; no doc required on test fns

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
strict_optional = true
warn_unused_ignores = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
warn_redundant_casts = true
warn_unused_configs = true
warn_no_return = true
warn_return_any = true
warn_unreachable = true
disallow_untyped_decorators = true
disallow_untyped_calls = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-m 'not integration' --cov=air_marshall --cov-report=term-missing --cov-report=xml"
asyncio_mode = "strict"
asyncio_default_fixture_loop_scope = "function"
filterwarnings = [
    "error",
    "ignore::DeprecationWarning:websockets",
    "ignore::DeprecationWarning:uvicorn",
]
markers = [
    "integration: marks tests requiring live hardware or external services (deselect with '-m not integration'). Tests requiring RPi hardware must skip gracefully when hardware is unavailable (e.g. pytest.importorskip or a custom skip fixture).",
]

[tool.coverage.run]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
